move portfolio and stock to their own file
